fixed issue where rating bar selection was not updating the score textview within rating dialog. Another issue fixed was the rating dialog wasn't sending the correct rate chosen, but instead was just sending a hard coded 0.0.

+++ app/src/main/java/movienight/javi/com/movienight/listeners/RateSelectedListener.java
@@ -6,5 +6,5 @@ package movienight.javi.com.movienight.listeners;
 
 public interface RateSelectedListener {
 
-    void onRateDone(Double rate);
+    void onRateDone(float rate);
 }
